+++ include/qristal/core/session_utils.hpp
@@ -78,7 +78,7 @@ namespace qristal
     /// [AER backend] Simulator type (e.g., state vec, density matrix, etc.)
     std::string aer_sim_type;
     /// [Emulator tensor network accelerators] Measurement sampling method
-    std::string measure_sample_sequential;
+    std::string measure_sample_method;
   };
 
 }